-Fixed diet loading from DB

+++ app/src/main/java/com/example/healthcared/Modelo/Comida.kt
@@ -5,7 +5,7 @@ package com.example.healthcared.Modelo
  * time indica a qué hora pertenece la comida
  * ingredients indica los ingredientes, puede ser null
  */
-class Comida(val foodName: String, val time: Int, val recipeLink: String, val dia: Int) {
+class Comida(val foodName: String, val time: Long, val recipeLink: String, val dia: Long) {
     override fun toString(): String {
         return super.toString()
     }
